+++ src/logger/logger.ts
@@ -1,10 +1,4 @@
 import pino, { type TransportTargetOptions } from "pino";
-import pinoCaller from "pino-caller";
-
-
-
-
-const isDevelopment = process.env.NODE_ENV === "development";
 
 
 
@@ -18,23 +12,42 @@ const developmentOnlyTargets: TransportTargetOptions[] = [
 ];
 
 
-const commonTargets: TransportTargetOptions[] = [
-    {
-        target: "pino/file",
-        level: "trace",
-        options: { destination: "./logs/log.txt" },
-    },
+const productionOnlyTargets: TransportTargetOptions[] = [
     {
         target: "pino/file",
         level: "warn",
-        options: { destination: "./logs/warnOnly.txt" },
+        options: { destination: 1 },
+    },
+];
+
+
+const commonTargets: TransportTargetOptions[] = [
+    {
+        target: "pino/file",
+        level: "warn",
+        options: {
+            destination: "./logs/errors_and_warnings.txt",
+            mkdir: true,
+        },
+    },
+    {
+        target: "pino/file",
+        level: "info",
+        options: {
+            destination: "./logs/info.txt",
+            mkdir: true,
+        },
     },
 ];
 
 
 
 
-const targets = isDevelopment ? [...developmentOnlyTargets, ...commonTargets] : commonTargets;
+const targets = process.env.NODE_ENV === "development" ? [...developmentOnlyTargets, ...commonTargets] : (
+    process.env.NODE_ENV === "production" ?
+        [...productionOnlyTargets, ...commonTargets] :
+        commonTargets
+);
 
 
 const transport = pino.transport({ targets });
@@ -42,4 +55,10 @@ const transport = pino.transport({ targets });
 
 
 
-export const logger = isDevelopment ? pinoCaller(pino(transport), { relativeTo: "resolver (webpack-internal:///)" }) : pino(transport);     //TODO: decide how to prettify stack trace better. probably stop using pinoCaller and make my own wrapper
+export const logger = pino(
+    {
+        base: undefined,
+        nestedKey: "payload",
+    },
+    transport
+);
